# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IAmSurfer API",
    "version": "1",
    "description": "API REST versionada (/api/v1) do app mobile IAmSurfer. Autenticação via JWT Bearer (access 15min + refresh 30d rotativo). Erros no formato { \"error\": { \"code\", \"message\", \"fields\" } }. Feeds paginados por cursor: { \"items\": [...], \"next_cursor\": ... }."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/me/trips": {
      "get": {
        "operationId": "get_api_trips_api_my_trips",
        "tags": [
          "trips"
        ],
        "summary": "GET /me/trips",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Erro no formato padrão da API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessária ou token inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/trips": {
      "get": {
        "operationId": "get_api_trips_api_list_trips",
        "tags": [
          "trips"
        ],
        "summary": "GET /trips",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Erro no formato padrão da API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post_api_trips_api_create_trip",
        "tags": [
          "trips"
        ],
        "summary": "POST /trips",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Erro no formato padrão da API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessária ou token inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/trips/{trip_id}/leave": {
      "post": {
        "operationId": "post_api_trips_api_leave_trip",
        "tags": [
          "trips"
        ],
        "summary": "POST /trips/{trip_id}/leave",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Erro no formato padrão da API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessária ou token inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/trips/{trip_id}/join": {
      "post": {
        "operationId": "post_api_trips_api_join_trip",
        "tags": [
          "trips"
        ],
        "summary": "POST /trips/{trip_id}/join",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Erro no formato padrão da API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessária ou token inválido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/trips/{trip_id}": {
      "get": {
        "operationId": "get_api_trips_api_get_trip",
        "tags": [
          "trips"
        ],
        "summary": "GET /trips/{trip_id}",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "default": {
            "description": "Erro no formato padrão da API.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "trip_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "Formato de erro único da API.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Código estável, legível por máquina."
              },
              "message": {
                "type": "string",
                "description": "Mensagem humana (pt-BR)."
              },
              "fields": {
                "type": "object",
                "description": "Erros de validação por campo.",
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    }
  }
}